before norminate

+++ Libft/ft_atoi.c
@@ -6,20 +6,19 @@
 /*   By: mhrima <mhrima@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/08 23:41:16 by mhrima            #+#    #+#             */
-/*   Updated: 2022/10/19 14:10:55 by mhrima           ###   ########.fr       */
+/*   Updated: 2023/02/14 02:15:49 by mhrima           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <limits.h>
 #include <stdlib.h>
 
-int	ft_isspace(char c)
+int ft_isspace(char c)
 {
-	return (c == '\t' || c == '\n' || c == '\v' || c == '\f' || c == '\r'
-		|| c == ' ');
+	return (c == '\t' || c == '\n' || c == '\v' || c == '\f' || c == '\r' || c == ' ');
 }
 
-int	the_sign(char c, int *i)
+int the_sign(char c, int *i)
 {
 	if (c == '+' || c == '-')
 	{
@@ -30,11 +29,11 @@ int	the_sign(char c, int *i)
 	return (1);
 }
 
-int	ft_atoi(const char *str)
+long int ft_atoi(const char *str)
 {
-	int			i;
-	int			sign;
-	long int	num;
+	int i;
+	int sign;
+	long int num;
 
 	i = 0;
 	num = 0;
@@ -44,13 +43,11 @@ int	ft_atoi(const char *str)
 	while (str[i] <= '9' && str[i] >= '0')
 	{
 		num = num * 10 + sign * (str[i] - '0');
-		if (sign == 1 && str[i + 1] <= '9' && str[i + 1] >= '0'
-			&& num > (LONG_MAX - sign * (str[i + 1] - '0')) / 10)
+		if (sign == 1 && str[i + 1] <= '9' && str[i + 1] >= '0' && num > (LONG_MAX - sign * (str[i + 1] - '0')) / 10)
 			return (-1);
-		if (sign == -1 && str[i + 1] <= '9' && str[i + 1] >= '0'
-			&& num < (LONG_MIN - sign * (str[i + 1] - '0')) / 10)
+		if (sign == -1 && str[i + 1] <= '9' && str[i + 1] >= '0' && num < (LONG_MIN - sign * (str[i + 1] - '0')) / 10)
 			return (0);
 		i++;
 	}
-	return ((int)num);
+	return (num);
 }
